Refactor slot tests
